chore: move Debug, Math, MaxMeasure to utils
