+++ src/base/utils/Math.ts
