+++ src/charts/column/ColumnChart.ts
@@ -3,10 +3,10 @@ import { ColumnAxis } from '../../base/Axis';
 import { Box } from '../../base/Box';
 import { Scenario, extractParallelValues } from '../../base/Data';
 import { ParallelDataType } from '../../base/DataTypes';
-import { round } from '../../base/Math';
-import { MaxMeasure } from '../../base/MaxMeasure';
 import { Rect } from '../../base/Rect';
 import { Text } from '../../base/Text';
+import { round } from '../../base/utils/Math';
+import { MaxMeasure } from '../../base/utils/MaxMeasure';
 import { Chart, ChartProps } from '../Chart';
 import { Constants } from '../Constants';
 import { SideExtension, SideExtensionProps } from './SideExtension';
